+++ tasks/task_3/simonyan_s_quicksort/quicksort.h
@@ -7,5 +7,4 @@
 
 void quickSortFunction(double* array, int left, int right);
 void Sort(double* arr, int n);
-
 #endif  // TASKS_TASK_3_SIMONYAN_S_QUICKSORT_QUICKSORT_H_
